+++ src/main/php/valuereader/CommonValueReader.php
@@ -179,6 +179,15 @@ interface CommonValueReader
     public function asDatespan(DatespanRange $range = null);
 
     /**
+     * read value as instance of given enum
+     *
+     * @param   string  $enumClass  name of enum class to derive value from
+     * @return  \stubbles\lang\Enum
+     * @since   5.0.0
+     */
+    public function asEnum($enumClass);
+
+    /**
      * returns value if it is an ip address, and null otherwise
      *
      * @return  string