+++ src/main/php/valuereader/DefaultValueReader.php
@@ -294,6 +294,19 @@ class DefaultValueReader implements CommonValueReader
     }
 
     /**
+     * read value as instance of given enum
+     *
+     * @param   string  $enumClass  name of enum class to derive value from
+     * @return  \stubbles\lang\Enum
+     * @since   5.0.0
+     */
+    public function asEnum($enumClass)
+    {
+        $this->checkDefaultType(function() use ($enumClass) { return $this->default instanceof $enumClass;}, $enumClass);
+        return $this->default;
+    }
+
+    /**
      * returns value if it is an ip address, and null otherwise
      *
      * @return  string