+++ src/main/php/valuereader/MissingValueReader.php
@@ -265,6 +265,21 @@ class MissingValueReader implements CommonValueReader
     }
 
     /**
+     * read value as instance of given enum
+     *
+     * @param   string  $enumClass  name of enum class to derive value from
+     * @return  \stubbles\lang\Enum
+     * @since   5.0.0
+     */
+    public function asEnum($enumClass)
+    {
+        $this->reportError(
+                'FIELD_EMPTY' === $this->defaultErrorId ? 'FIELD_NO_SELECT' : $this->defaultErrorId
+        );
+        return null;
+    }
+
+    /**
      * returns value if it is an ip address, and null otherwise
      *
      * @return  string